backend/requirements.txt
Django==5.1.6

# Django REST Framework
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1

# CORS
django-cors-headers==4.3.1

# Auth (optional but fine)
django-allauth==0.57.0

# Environment Variables
python-decouple==3.8

pdfplumber==0.10.3
reportlab==4.0.7

opencv-python-headless==4.8.1.78
pytesseract==0.3.10
Pillow>=10.3.0


requests==2.31.0

gunicorn==21.2.0
whitenoise==6.6.0

psycopg2-binary==2.9.9

numpy==1.26.4
